requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.11.4
google-auth==2.23.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
apify-client==1.6.0
aiohttp==3.8.5